+++ src/frontend/components/VideoPlayer/index.tsx
@@ -1,5 +1,6 @@
-import { Box } from 'grommet';
-import React, { useEffect, useRef, useState } from 'react';
+import { Box, Text } from 'grommet';
+import React, { useRef, useState } from 'react';
+import { defineMessages, useIntl } from 'react-intl';
 import { Redirect } from 'react-router';
 
 import { useThumbnail } from '../../data/stores/useThumbnail';
@@ -15,11 +16,20 @@ import {
   videoSize,
 } from '../../types/tracks';
 import { VideoPlayerInterface } from '../../types/VideoPlayer';
+import { useAsyncEffect } from '../../utils/useAsyncEffect';
 import { Maybe, Nullable } from '../../utils/types';
 import { Chat } from '../Chat';
 import { DownloadVideo } from '../DownloadVideo';
 import { FULL_SCREEN_ERROR_ROUTE } from '../ErrorComponents/route';
 import { Transcripts } from '../Transcripts';
+
+const messages = defineMessages({
+  liveOnGoing: {
+    defaultMessage: 'Live is starting and will be displayed soon.',
+    description: 'Waiting message when a live is starting',
+    id: 'components.VideoPlayer.LiveOnDoing',
+  },
+});
 
 const trackTextKind: { [key in timedTextMode]?: string } = {
   [timedTextMode.CLOSED_CAPTIONING]: 'captions',
@@ -35,6 +45,7 @@ const VideoPlayer = ({
   video: baseVideo,
   playerType,
 }: BaseVideoPlayerProps) => {
+  const intl = useIntl();
   const [player, setPlayer] = useState(
     undefined as Maybe<VideoPlayerInterface>,
   );
@@ -69,13 +80,13 @@ const VideoPlayer = ({
    * Initialize the video player.
    * Noop out if the video is missing, render will redirect to an error page.
    */
-  useEffect(() => {
+  useAsyncEffect(async () => {
     getChoices();
 
     if (video) {
       // Instantiate the player and keep the instance in state
       setPlayer(
-        createPlayer(
+        await createPlayer(
           playerType,
           videoNodeRef.current!,
           setPlayerCurrentTime,
@@ -155,9 +166,14 @@ const VideoPlayer = ({
             />
           ))}
       </video>
-      {video.show_download && <DownloadVideo urls={video.urls!} />}
-      {transcripts.length > 0 && (
+      {player && video.show_download && <DownloadVideo urls={video.urls!} />}
+      {player && transcripts.length > 0 && (
         <Transcripts transcripts={transcripts as TimedTextTranscript[]} />
+      )}
+      {!player && video.live_state && (
+        <Text size="large" textAlign="center">
+          {intl.formatMessage(messages.liveOnGoing)}
+        </Text>
       )}
       {video.live_state !== null && video.xmpp && <Chat video={video} />}
     </Box>